Derive Macro mimicry::CallReal

#[derive(CallReal)]
{
    // Attributes available to this derive:
    #[mock]
}

Derives the CallReal trait for a struct allowing to switch to real implementations for partial mocking or spying.

Field attributes

Field attributes are placed in a #[mock(...)] attribute on a struct / enum.

switch

Indicates that a field is a RealCallSwitch. This is usually detected automatically by the field type, so an explicit declaration is reserved for extraordinary cases. Specified as #[mock(switch)].